add local spatial measures; refactor spatial measures
